graph, chain, core: added a BlockStream trait and turned existing one into PollingBlockStream impl
This is another small PR for upcoming Firehose feature. The BlockStream trait will enable using a PollingBlockStream (actual implementation) and an upcoming FirehoseBlockStream (next or so PR in line).

We decided on submitting a PR with just that for now to have few focused PRs around.
